+++ packages/flow-form-builder/CHANGELOG.md
@@ -1,7 +1,12 @@
 <h4 class="margin-btm-8">Release Notes</h4>
 
-
 # Change Log
+
+## [2.0.4] - 2023-10-16
+
+### Patch Changes
+
+- v1.8.0 and v1.8.1 commits merged.
 
 ## [2.0.3] - 2023-10-12
 
@@ -11,7 +16,6 @@
 - Updated dependencies
 - `@cldcvr/flow-core-config@1.1.3`
 - `@cldcvr/flow-core@2.0.3`
-
 
 ## [2.0.2] - 2023-10-10
 
@@ -43,6 +47,18 @@
 - `@cldcvr/flow-core@2.0.0`
 <hr class="margin-btm-32" />
 <p class="margin-btm-24">All notable changes to this project will be documented in this file. See <a>Conventional Commits</a> for commit guidelines. </p>
+
+## [1.8.1] - 2023-09-13
+
+### Improvements
+
+- Array field description font size fixed.
+
+## [1.8.0] - 2023-09-13
+
+### Features
+
+- Custom validation message using callback function for dynamic messages.
 
 ## [1.7.8] - 2023-09-08